requirements.txt
boto3
botocore
streamlit
nltk
opencv-python
Pillow
python-dotenv
matplotlib